requirements.txt
fastapi
uvicorn
pydantic
pytest
pytest-cov
bcrypt
python-jose
passlib[bcrypt]
requests
